test(shields): split audit reader coverage
